Move `fn basic()`/`fn required()` from `GenericMediaTrackConstraints<_>` to `GenericMandatoryMediaTrackConstraints<_>` and add tests
